# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PUFFIN SRL - Plataforma Integral de Gestión Operativa
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Iniciar sesión
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SesionData"
        "401":
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Obtener usuario actual
      responses:
        "200":
          description: Usuario actual
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
        "401":
          description: No autorizado
...
  /empleados:
    get:
      operationId: getEmpleados
      tags: [empleados]
      summary: Listar operarios
      parameters:
        - name: estado
          in: query
          schema:
            type: string
            enum: [activo, inactivo]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de operarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Empleado"
    post:
      operationId: createEmpleado
      tags: [empleados]
      summary: Crear operario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmpleadoInput"
      responses:
        "201":
          description: Operario creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Empleado"
  /empleados/{id}:
    get:
      operationId: getEmpleado
      tags: [empleados]
      summary: Obtener operario por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Operario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Empleado"
    put:
      operationId: updateEmpleado
      tags: [empleados]
      summary: Actualizar operario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmpleadoUpdate"
      responses:
        "200":
          description: Operario actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Empleado"
...
  /combustible:
    get:
      operationId: getCombustible
      tags: [combustible]
      summary: Listar registros de combustible
      parameters:
        - name: maquina_id
          in: query
          schema:
            type: integer
        - name: empleado_id
          in: query
          schema:
            type: integer
        - name: fecha_desde
          in: query
          schema:
            type: string
        - name: fecha_hasta
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Registros de combustible
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RegistroCombustible"
    post:
      operationId: createCombustible
      tags: [combustible]
      summary: Registrar carga de combustible
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CombustibleInput"
      responses:
        "201":
          description: Registro creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistroCombustible"
...
  /documentos:
    get:
      operationId: getDocumentos
      tags: [documentos]
      summary: Listar documentos
      parameters:
        - name: tipo
          in: query
          schema:
            type: string
        - name: proximo_vencimiento
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Lista de documentos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Documento"
    post:
      operationId: createDocumento
      tags: [documentos]
      summary: Registrar documento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentoInput"
      responses:
        "201":
          description: Documento creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Documento"
  /alertas:
    get:
      operationId: getAlertas
      tags: [alertas]
      summary: Listar alertas
      parameters:
        - name: prioridad
          in: query
          schema:
            type: string
            enum: [roja, amarilla, azul]
        - name: estado
          in: query
          schema:
            type: string
            enum: [activa, resuelta]
      responses:
        "200":
          description: Lista de alertas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alerta"
  /alertas/{id}:
    put:
      operationId: updateAlerta
      tags: [alertas]
      summary: Actualizar alerta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertaUpdate"
      responses:
        "200":
          description: Alerta actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alerta"
  /incidentes:
    get:
      operationId: getIncidentes
      tags: [incidentes]
      summary: Listar incidentes
      parameters:
        - name: maquina_id
          in: query
          schema:
            type: integer
        - name: empleado_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Lista de incidentes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Incidente"
    post:
      operationId: createIncidente
      tags: [incidentes]
      summary: Registrar incidente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidenteInput"
      responses:
        "201":
          description: Incidente registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incidente"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginInput:
      type: object
      required: [usuario, pin]
      properties:
        usuario:
          type: string
        pin:
          type: string
    SesionData:
      type: object
      required: [token, usuario]
      properties:
        token:
          type: string
        refreshToken:
          type: string
        usuario:
          $ref: "#/components/schemas/Usuario"
    Usuario:
      type: object
      required: [id, nombre, apellido, usuario, rol]
      properties:
        id:
          type: integer
        empresa_id:
          type: integer
        nombre:
          type: string
        apellido:
          type: string
        usuario:
          type: string
        rol:
          type: string
          enum: [administrador, empleado]
        activo:
          type: boolean
...
    Empleado:
      type: object
      required: [id, nombre, apellido, estado]
      properties:
        id:
          type: integer
        empresa_id:
          type: integer
        nombre:
          type: string
        apellido:
          type: string
        dni:
          type: string
        telefono:
          type: ["string", "null"]
        cargo:
          type: ["string", "null"]
        estado:
          type: string
          enum: [activo, inactivo, licencia]
        fecha_ingreso:
          type: ["string", "null"]
        jornada_activa:
          type: boolean
        alertas_count:
          type: integer
    EmpleadoInput:
      type: object
      required: [nombre, apellido, dni]
      properties:
        nombre:
          type: string
        apellido:
          type: string
        dni:
          type: string
        telefono:
          type: string
        cargo:
          type: string
        fecha_ingreso:
          type: string
    EmpleadoUpdate:
      type: object
      properties:
        nombre:
          type: string
        apellido:
          type: string
        telefono:
          type: string
        cargo:
          type: string
        estado:
          type: string
          enum: [activo, inactivo, licencia]
...
    RegistroCombustible:
      type: object
      required: [id, maquina_id, empleado_id, fecha, litros]
      properties:
        id:
          type: integer
        empresa_id:
          type: integer
        empleado_id:
          type: integer
        maquina_id:
          type: integer
        empleado_nombre:
          type: string
        maquina_nombre:
          type: string
        fecha:
          type: string
        litros:
          type: number
        precio:
          type: ["number", "null"]
        importe:
          type: ["number", "null"]
        estacion:
          type: ["string", "null"]
        kilometraje:
          type: ["number", "null"]
    CombustibleInput:
      type: object
      required: [maquina_id, empleado_id, litros]
      properties:
        maquina_id:
          type: integer
        empleado_id:
          type: integer
        litros:
          type: number
        precio:
          type: number
        importe:
          type: number
        estacion:
          type: string
        kilometraje:
          type: number
        ubicacion:
          type: string
...
    Documento:
      type: object
      required: [id, tipo, fecha_vencimiento]
      properties:
        id:
          type: integer
        empresa_id:
          type: integer
        tipo:
          type: string
        descripcion:
          type: ["string", "null"]
        entidad_tipo:
          type: ["string", "null"]
        entidad_id:
          type: ["integer", "null"]
        entidad_nombre:
          type: ["string", "null"]
        fecha_vencimiento:
          type: string
        archivo_url:
          type: ["string", "null"]
        estado:
          type: string
          enum: [vigente, proximo_vencimiento, vencido]
        dias_restantes:
          type: integer
    DocumentoInput:
      type: object
      required: [tipo, fecha_vencimiento]
      properties:
        tipo:
          type: string
        descripcion:
          type: string
        entidad_tipo:
          type: string
        entidad_id:
          type: integer
        fecha_vencimiento:
          type: string
    Alerta:
      type: object
      required: [id, tipo, prioridad, descripcion, estado, fecha]
      properties:
        id:
          type: integer
        empresa_id:
          type: integer
        tipo:
          type: string
        prioridad:
          type: string
          enum: [roja, amarilla, azul]
        descripcion:
          type: string
        estado:
          type: string
          enum: [activa, resuelta, ignorada]
        fecha:
          type: string
        entidad_tipo:
          type: ["string", "null"]
        entidad_id:
          type: ["integer", "null"]
        entidad_nombre:
          type: ["string", "null"]
    AlertaUpdate:
      type: object
      properties:
        estado:
          type: string
          enum: [activa, resuelta, ignorada]
    Incidente:
      type: object
      required: [id, descripcion, fecha]
      properties:
        id:
          type: integer
        empresa_id:
          type: integer
        empleado_id:
          type: ["integer", "null"]
        maquina_id:
          type: ["integer", "null"]
        empleado_nombre:
          type: ["string", "null"]
        maquina_nombre:
          type: ["string", "null"]
        tipo:
          type: string
          enum: [rotura, golpe, accidente, falla, problema_mecanico, otro]
        descripcion:
          type: string
        foto_url:
          type: ["string", "null"]
        fecha:
          type: string
        estado:
          type: string
          enum: [activo, resuelto]
    IncidenteInput:
      type: object
      required: [tipo, descripcion]
      properties:
        empleado_id:
          type: integer
        maquina_id:
          type: integer
        tipo:
          type: string
          enum: [rotura, golpe, accidente, falla, problema_mecanico, otro]
        descripcion:
          type: string
...