# lib/api-zod/src/generated/types/empleado.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PUFFIN SRL - Plataforma Integral de Gestión Operativa
 * OpenAPI spec version: 0.1.0
 */
import type { EmpleadoEstado } from './empleadoEstado';

export interface Empleado {
  id: number;
  empresa_id?: number;
  nombre: string;
  apellido: string;
  dni?: string;
  /** @nullable */
  telefono?: string | null;
  /** @nullable */
  cargo?: string | null;
  estado: EmpleadoEstado;
  /** @nullable */
  fecha_ingreso?: string | null;
  jornada_activa?: boolean;
  alertas_count?: number;
}