# lib/api-zod/src/generated/types/empleadoInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PUFFIN SRL - Plataforma Integral de Gestión Operativa
 * OpenAPI spec version: 0.1.0
 */

export interface EmpleadoInput {
  nombre: string;
  apellido: string;
  dni: string;
  telefono?: string;
  cargo?: string;
  fecha_ingreso?: string;
}