# lib/api-zod/src/generated/types/empleadoUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * PUFFIN SRL - Plataforma Integral de Gestión Operativa
 * OpenAPI spec version: 0.1.0
 */
import type { EmpleadoUpdateEstado } from './empleadoUpdateEstado';

export interface EmpleadoUpdate {
  nombre?: string;
  apellido?: string;
  telefono?: string;
  cargo?: string;
  estado?: EmpleadoUpdateEstado;
}